requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
Pillow>=10.1.0
numpy>=1.26.0
mediapipe>=0.10.8
httpx>=0.25.0
python-dotenv>=1.0.0
aiofiles>=23.2.0
pydantic>=2.5.0
